Move common mapping functions to shared mapping component
